fix: ensure DB migration runs (#1120)
* ensure DB migration runs

* add beta to branch QA
