+++ .github/workflows/branch--lint-unit-and-smoke-test.yml
@@ -3,6 +3,7 @@ on:
   pull_request:
     branches:
       - main
+      - beta
     paths-ignore:
       - "docs/**"
       - "**/README.md"
